soc/amd/*/root_complex: use unsigned long for resource index
Use an unsigned long as resource index type instead of an int to match
the data type used for the index in the resource struct.

Reviewed-on: https://review.coreboot.org/c/coreboot/+/80267
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/soc/amd/common/block/include/amdblocks/root_complex.h
@@ -27,6 +27,6 @@ const struct non_pci_mmio_reg *get_iohc_non_pci_mmio_regs(size_t *count);
 
 signed int get_iohc_fabric_id(struct device *domain);
 
-void read_fsp_resources(struct device *dev, unsigned int *idx);
+void read_fsp_resources(struct device *dev, unsigned long *idx);
 
 #endif /* AMD_BLOCK_ROOT_COMPLEX_H */
